requirements.txt
aioredis==2.0.1
faiss-cpu==1.9.0
fastapi==0.115.0
httpx==0.27.2
huggingface_hub
jinja2
langchain
langchain-core
langchain_huggingface
langgraph
numpy==1.26.4
opencv-python
pdf2image
pdfminer.six==20231228
Pillow==10.4.0
pytesseract==0.3.13
pydantic==2.9.2
pydantic-settings==2.5.2
pyyaml
python-dotenv
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
pytest
redis==5.2.0
requests
sentence-transformers==3.2.1
uvicorn[standard]==0.32.0

# PDF Generation Libraries
reportlab==4.2.5
PyPDF2==3.0.1